Added naive implementation of information search engine
This version uses lists both as information base and informations. It has no
optimization whatsoever as far as matching the request against the information
base is concerned.

The functions are EVERY-APPLY, SOME-APPLY, CLAUSE-PREDICATE and
REQUEST-PREDICATE.

Methods on informations as lists are INFORMATION-SYMBOL and
INFORMATION-ARGUMENTS.

Methods on information bases as lists are INFORMATION-SEARCH.